Added SmartCopy
It should be good, I hope it's good. I don't know how to write an unit test for it, I wish I did. I think this will be easier to write unit tests when I implement the MemoryFileAccessor, still a work in progress (like.. everything in the life and universe, perhaps. NVM)