Added the ability for BDadosTransactions to auto-commit/rollback
Added the functions Throw and Complete to fine-tune Commit/Rollback decision in user-code. User can still manually commit or rollback and that will cause the auto to not do anything. In general auto-committing should only fail if you get an exception in the middle of AsyncEnumerable operation (will try to commit), shouldn't be bad for read-only operations, but I strongly suggest using try-catch in IAsyncEnumerations and throwing with the tsn.Throw() so that the dispose function rolls back instead of committing